requirements.txt
# ═══════════════════════════════════════════════════════════
# Video Localization Pipeline — Python Dependencies
# ═══════════════════════════════════════════════════════════

# Core AI
whisperx @ git+https://github.com/m-bain/whisperX.git
demucs
google-generativeai>=0.5.0

# Audio processing
pydub>=0.25.1
pysubs2>=1.7.0

# Video download
yt-dlp>=2024.1.0

# Async HTTP for TTS
aiohttp>=3.9.0
aiofiles>=23.0.0

# Google Cloud TTS
google-cloud-texttospeech>=2.16.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
